@@ -16,8 +16,7 @@ final class RegisterController extends AbstractController
 {
     use RegistersUsers;
 
-    /** @var string */
-    protected $redirectTo = RouteServiceProvider::HOME;
+    protected string $redirectTo = RouteServiceProvider::HOME;
 
     public function __construct()
     {
