@@ -8,8 +8,8 @@ use Illuminate\Support\Facades\Route;
 
 // API Routes
 
-Route::middleware('auth:api')->get('/user', function (Request $request) {
-    return $request->user();
-});
+Route::middleware('auth:api')->get('/user',
+    fn(Request $request) => $request->user()
+);
 
 Route::post('/upload-files', [TestController::class, 'uploadFiles']);
