@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Http\Controllers;
 
 use App\Events\TestEvent;
+use Illuminate\Contracts\Config\Repository as Config;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 
@@ -17,9 +18,11 @@ final class TestController
 
     public function testValue(): string
     {
-        return 'Test value is: ' . config('test_value');
+        $config = app()->get(Config::class);
+        return 'Test value is: ' . $config->get('test_value');
     }
 
+    /** @return array<string, array<bool|int>> */
     public function uploadFiles(Request $request): array
     {
         $response = [];
