@@ -10,16 +10,14 @@ use Illuminate\Http\Request;
 
 final class Authenticate extends Middleware
 {
-    /**
-     * @param Request $request
-     * @return string|null
-     */
+    /** @param Request $request */
     protected function redirectTo($request): ?string
     {
         $urlGenerator = app()->get(UrlGenerator::class);
         if (! $request->expectsJson()) {
             $urlGenerator->route('login');
         }
+
         return null;
     }
 }
