@@ -41,7 +41,8 @@ use Illuminate\Notifications\Notifiable;
  */
 class User extends BaseUser
 {
-    use HasFactory, Notifiable;
+    use HasFactory;
+    use Notifiable;
 
     /** @var array */
     protected $fillable = [
@@ -66,7 +67,7 @@ class User extends BaseUser
         return $this->name;
     }
 
-    public function setName(string $name): User
+    public function setName(string $name): self
     {
         $this->name = $name;
         return $this;
@@ -77,7 +78,7 @@ class User extends BaseUser
         return $this->email;
     }
 
-    public function setEmail(string $email): User
+    public function setEmail(string $email): self
     {
         $this->email = $email;
         return $this;
@@ -88,7 +89,7 @@ class User extends BaseUser
         return $this->password;
     }
 
-    public function setPassword(string $password): User
+    public function setPassword(string $password): self
     {
         $this->password = $password;
         return $this;
@@ -99,7 +100,7 @@ class User extends BaseUser
         return $this->email_verified_at;
     }
 
-    public function setEmailVerifiedAt(Carbon $verifiedDate): User
+    public function setEmailVerifiedAt(Carbon $verifiedDate): self
     {
         $this->email_verified_at = $verifiedDate;
         return $this;
