@@ -8,7 +8,7 @@ use Illuminate\Cache\RateLimiter;
 use Illuminate\Cache\RateLimiting\Limit;
 use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
+use Illuminate\Routing\Router;
 use function base_path;
 use function optional;
 
@@ -24,13 +24,14 @@ final class RouteServiceProvider extends ServiceProvider
     {
         $this->configureRateLimiting();
 
-        $this->routes(function (): void {
-            Route::prefix('api')
+        $router = app()->get(Router::class);
+        $this->routes(function() use ($router): void {
+            $router->prefix('api')
                 ->middleware('api')
                 ->namespace($this->namespace)
                 ->group(base_path('routes/api.php'));
 
-            Route::middleware('web')
+            $router->middleware('web')
                 ->namespace($this->namespace)
                 ->group(base_path('routes/web.php'));
         });
@@ -39,10 +40,8 @@ final class RouteServiceProvider extends ServiceProvider
     protected function configureRateLimiting(): void
     {
         $rateLimiter = app()->get(RateLimiter::class);
-        $rateLimiter->for('api', function (Request $request): Limit {
-            return Limit::perMinute(60)->by(
-                (string) optional($request->user())->id ?: $request->ip()
-            );
-        });
+        $rateLimiter->for('api', fn(Request $request): Limit => Limit::perMinute(60)->by(
+            (string) optional($request->user())->id ?: $request->ip()
+        ));
     }
 }
